refactor: i18n-page-edit
