+++ shell/app/i18n-page-edit/config.ts
