+++ shell/app/i18n.ts
@@ -12,7 +12,7 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import i18n, { getCurrentLocale, getLang, isZh, setLocale } from 'core/i18n';
-import i18nWrapper from './i18n_rewrite/overwrite_i18n';
+import i18nWrapper from './i18n-page-edit/overwrite_i18n';
 import zh from './locales/zh.json';
 import en from './locales/en.json';
 import defaultZh from '../../locales/zh.json';
